refactor: extract active workers
