+++ apps/worker/src/app/shared/shared.module.ts
@@ -1,4 +1,4 @@
-import { Module, Provider } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import {
   ChangeRepository,
   DalService,
@@ -60,8 +60,7 @@ import * as packageJson from '../../../package.json';
 import { CreateLog } from './logs';
 import { JobTopicNameEnum } from '@novu/shared';
 import { ActiveJobsMetricService, ExecutionLogWorker, StandardWorker, WorkflowWorker } from '../workflow/services';
-import { SubscriberProcessWorker } from '../workflow/services/subscriber-process.worker';
-import { WorkflowModule } from '../workflow/workflow.module';
+import { queuesToProcess, UNIQUE_QUEUE_DEPENDENCIES } from '../../config/worker-init.config';
 
 const DAL_MODELS = [
   UserRepository,
@@ -130,54 +129,6 @@ const PROVIDERS = [
   ...DAL_MODELS,
   ActiveJobsMetricService,
 ];
-
-const validQueueEntries = Object.keys(JobTopicNameEnum).map((key) => JobTopicNameEnum[key]);
-const queuesToProcess =
-  process.env.WORKER_QUEUES?.split(',').map((queue) => {
-    const queueName = queue.trim();
-    if (!validQueueEntries.includes(queueName)) {
-      throw new Error(`Invalid queue name ${queueName}`);
-    }
-
-    return queueName;
-  }) || [];
-
-export const ACTIVE_WORKERS: Provider[] | any[] = [];
-
-const WORKER_MAPPING = {
-  [JobTopicNameEnum.STANDARD]: {
-    workerClass: StandardWorker,
-    queueDependencies: [JobTopicNameEnum.EXECUTION_LOG, JobTopicNameEnum.WEB_SOCKETS, JobTopicNameEnum.STANDARD],
-  },
-  [JobTopicNameEnum.WORKFLOW]: {
-    workerClass: WorkflowWorker,
-    queueDependencies: [JobTopicNameEnum.EXECUTION_LOG, JobTopicNameEnum.PROCESS_SUBSCRIBER],
-  },
-  [JobTopicNameEnum.EXECUTION_LOG]: {
-    workerClass: ExecutionLogWorker,
-    queueDependencies: [],
-  },
-  [JobTopicNameEnum.PROCESS_SUBSCRIBER]: {
-    workerClass: SubscriberProcessWorker,
-    queueDependencies: [JobTopicNameEnum.EXECUTION_LOG],
-  },
-};
-
-const QUEUE_DEPENDENCIES = queuesToProcess.reduce((history, queue) => {
-  const queueDependencies = WORKER_MAPPING[queue]?.queueDependencies || [];
-
-  return [...history, ...queueDependencies];
-}, []);
-
-const UNIQUE_QUEUE_DEPENDENCIES = [...new Set(QUEUE_DEPENDENCIES)];
-
-if (!queuesToProcess.length) {
-  ACTIVE_WORKERS.push(StandardWorker, WorkflowWorker, ExecutionLogWorker, SubscriberProcessWorker);
-} else {
-  queuesToProcess.forEach((queue) => {
-    ACTIVE_WORKERS.push(WORKER_MAPPING[queue].workerClass);
-  });
-}
 
 @Module({
   imports: [
